fix gotNewComment on specific system_event type

+++ Qiscus/Qiscus/Qiscus.swift
@@ -975,42 +975,41 @@ extension Qiscus:CocoaMQTTDelegate{
                         let commentType = json["type"].stringValue
                         if commentType == "system_event" {
                             let payload = json["payload"]
-                            if payload["object_email"].stringValue == QiscusMe.sharedInstance.email {
-                                switch payload["type"].stringValue {
-                                case "remove_member":
+                            let type = payload["type"].stringValue
+                            switch type {
+                            case "remove_member", "add_member" :
+                                if payload["object_email"].stringValue == QiscusMe.sharedInstance.email {
                                     DispatchQueue.main.async {autoreleasepool{
-                                        if commentId > QiscusMe.sharedInstance.lastCommentId{
-                                            if let roomDelegate = QiscusCommentClient.shared.roomDelegate {
-                                                let comment = QComment.tempComment(fromJSON: json)
-                                                roomDelegate.gotNewComment(comment)
-                                            }
-                                            QiscusMe.updateLastCommentId(commentId: commentId)
+                                        if let roomDelegate = QiscusCommentClient.shared.roomDelegate {
+                                            let comment = QComment.tempComment(fromJSON: json)
+                                            roomDelegate.gotNewComment(comment)
                                         }
-                                        if let chatView = Qiscus.shared.chatViews[roomId] {
-                                            if chatView.isPresence {
-                                                chatView.goBack()
+                                        if type == "remove_member" {
+                                            if let chatView = Qiscus.shared.chatViews[roomId] {
+                                                if chatView.isPresence {
+                                                    chatView.goBack()
+                                                }
+                                                Qiscus.shared.chatViews[roomId] = nil
                                             }
-                                            Qiscus.shared.chatViews[roomId] = nil
-                                        }
-                                        
-                                        if let room = QRoom.room(withId: roomId){
                                             Qiscus.chatRooms[roomId] = nil
-                                            QRoom.deleteRoom(room: room)
-                                        }
-                                        }}
-                                break
-                                default:
-                                    DispatchQueue.main.async {autoreleasepool{
-                                        if commentId > QiscusMe.sharedInstance.lastCommentId{
-                                            if let roomDelegate = QiscusCommentClient.shared.roomDelegate {
-                                                let comment = QComment.tempComment(fromJSON: json)
-                                                roomDelegate.gotNewComment(comment)
+                                            if let room = QRoom.room(withId: roomId){
+                                                QRoom.deleteRoom(room: room)
                                             }
                                             QiscusMe.updateLastCommentId(commentId: commentId)
                                         }
                                     }}
-                                break
                                 }
+                            break
+                            case "create_room", "join_room", "left_room", "change_room_name", "change_room_avatar", "custom":
+                                DispatchQueue.main.async {
+                                    if let roomDelegate = QiscusCommentClient.shared.roomDelegate {
+                                        let comment = QComment.tempComment(fromJSON: json)
+                                        roomDelegate.gotNewComment(comment)
+                                    }
+                                }
+                            break
+                            default:
+                            break
                             }
                         }
                     }else{
